Fixed "Page stuck when switching issue"
